Add gurobi_solver_internal.h/cc
